MGComics/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==4.0.3
django-ckeditor==6.4.0
django-cleanup==6.0.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
